Add files via upload
I made a BASH script on Ubuntu 20.04 that uses Squahfs or Tar to create a backup of the root directory to automate this process. This script creates a log, cleans up old backups automatically, and generates a Squashfs or Tar (or both) backup in specified directories. This script must be run as sudo or root. This script can be run from any directory with the correct permissions. This script is GNU so feel free to download or modify.